Combine Python and Rust library source trees
Instead of one Rust package per tool, I want to have one Rust package
with many tools. This allows more consistent tooling and tests, and
allows sharing code between them.

I also want the project as a whole to feel more consistent, despite
containing Python, C++, and Rust. Eventually I'd like to replace the
geom2graph tool with a better Python/Rust replacement as it's quite
disappointing to me in its current state, and I'd prefer to not write
more C++ ;)
