--- tools/dla/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "dla"
-version = "0.1.0"
-authors = ["Austin Gill <Notgnoshi@gmail.com>"]
-edition = "2018"
-description = "Off-lattice diffusion limited aggregation"
-
-[dependencies]
-kdtree = "0.6"
-petgraph = "0.5"
-rand = "0.8"
-
-log = "0.4"
-stderrlog = "0.5"
-structopt = "0.3"
