+++ tools/dla/main.rs
@@ -1,19 +1,20 @@
+use generative::dla::format_tgf;
+use generative::dla::format_wkt;
+use generative::dla::Model;
 use log::trace;
+
 mod cmdline;
-mod dla;
-mod wkio;
 
 fn main() {
     let args = cmdline::Options::from_args();
 
     stderrlog::new()
-        .module(module_path!())
         .quiet(args.quiet)
         .verbosity(args.verbose + 2) // Default to INFO level
         .init()
         .unwrap();
 
-    let mut model = dla::Model::new(
+    let mut model = Model::new(
         args.dimensions,
         // TODO: Seed type.
         args.seeds,
@@ -32,10 +33,10 @@ fn main() {
     let mut writer = args.get_output_writer();
     match args.format {
         cmdline::OutputFormat::GraphTGF => {
-            wkio::format_tgf(&mut writer, model.particle_graph);
+            format_tgf(&mut writer, model.particle_graph);
         }
         cmdline::OutputFormat::PointCloudWKT => {
-            wkio::format_wkt(&mut writer, model.particle_graph);
+            format_wkt(&mut writer, model.particle_graph);
         }
     };
 }
